flask_requirements.txt
redis
gevent
gevent-websocket
gevent-socketio
Flask
Flask-Redis